feat: AI設定を管理者ダッシュボードに移動
- /settings/ai を削除し、/admin/ai として管理者ダッシュボード配下に移動
- 管理者サイドバーに「AI 設定」リンクを追加
- 設定はシステム全体共通のため SuperAdmin 専用管理画面が適切
